+++ src/qbus/responsiblemodule.cpp
@@ -1,7 +1,7 @@
 #include "qbus.h"
 #include "trace/trace.h"
 
-BusDevice *Qbus::responsibleModule (u16 address)
+BusDevice *Qbus::responsibleModule (u32 address)
 {
 	for (size_t index = 0; index < LSI11_SIZE; ++index)
 	{
