+++ src/qbus/writebyte.cpp
@@ -1,11 +1,11 @@
 #include "qbus.h"
 #include "trace/trace.h"
 
-bool Qbus::writeByte (u16 address, u8 value)
+bool Qbus::writeByte (u32 address, u8 value)
 {
 	BusDevice *module;
 		
-	if ((module = responsibleModule(address)) != nullptr)
+	if ((module = responsibleModule (address)) != nullptr)
 	{
 		trace.bus (BusRecordType::Write, address, value);
 		module->writeByte (address, value);
